+++ gcloud/datastore/test___init__.py
@@ -13,7 +13,7 @@ class Test_get_connection(unittest2.TestCase):
         from gcloud.datastore import SCOPE
         from gcloud.datastore.connection import Connection
         from gcloud.test_credentials import _Client
-        from gcloud.test_credentials import _Monkey
+        from gcloud._testing import _Monkey
 
         CLIENT_EMAIL = 'phred@example.com'
         PRIVATE_KEY = 'SEEkR1t'
@@ -45,7 +45,7 @@ class Test_get_dataset(unittest2.TestCase):
         from gcloud.datastore.connection import Connection
         from gcloud.datastore.dataset import Dataset
         from gcloud.test_credentials import _Client
-        from gcloud.test_credentials import _Monkey
+        from gcloud._testing import _Monkey
 
         CLIENT_EMAIL = 'phred@example.com'
         PRIVATE_KEY = 'SEEkR1t'
