+++ gcloud/test_credentials.py
@@ -6,6 +6,7 @@ class TestCredentials(unittest2.TestCase):
     def test_get_for_service_account_wo_scope(self):
         from tempfile import NamedTemporaryFile
         from gcloud import credentials
+        from gcloud._testing import _Monkey
         CLIENT_EMAIL = 'phred@example.com'
         PRIVATE_KEY = 'SEEkR1t'
         client = _Client()
@@ -25,6 +26,7 @@ class TestCredentials(unittest2.TestCase):
     def test_get_for_service_account_w_scope(self):
         from tempfile import NamedTemporaryFile
         from gcloud import credentials
+        from gcloud._testing import _Monkey
         CLIENT_EMAIL = 'phred@example.com'
         PRIVATE_KEY = 'SEEkR1t'
         SCOPE = 'SCOPE'
@@ -51,20 +53,3 @@ class _Client(object):
     def SignedJwtAssertionCredentials(self, **kw):
         self._called_with = kw
         return self._signed
-
-
-class _Monkey(object):
-    # context-manager for replacing module names in the scope of a test.
-
-    def __init__(self, module, **kw):
-        self.module = module
-        self.to_restore = dict([(key, getattr(module, key)) for key in kw])
-        for key, value in kw.items():
-            setattr(module, key, value)
-
-    def __enter__(self):
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        for key, value in self.to_restore.items():
-            setattr(self.module, key, value)
